Fix vLLM inference spec and chat client
Omit the previous command ("binary" in the spec) as the vLLM images use the
entrypoint to run vLLM.

Fix --served-model-name option.
This refered to a non-existant model.model_name ctx variable.
Use model.alias instead.

Use precisely the same model name from the chat client.
vLLM will exit when this is not equal to the served-model-name.

Signed-off-by: Oliver Walsh <owalsh@redhat.com>